+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Updated storage parsing to use kva instead of kvar for qs_lb and qs_ub (storage in dss is defined to be able to handle any reactive power up to the thermal limit of the inverter)
 - Fixed `make_lossless!` to adjust how switches are made lossless
 - Fixed voltage sqr variable start values
 - Fixed p/q variable start values